tools/resources/requirements.txt
argparse==1.4.0
pyyaml==6.0.1
